Added a basic rgbLedsTryShiftNext(r, g, b) function which populates the circular buffer for the UART.
Unfortunately the UART requires bit reversal (it's LSb-first as opposed to SPI / WS2812B that are MSb-first).  Bit reversal is cycle-hungry when there are only 80 cycles between UART bytes to play with (SPI timing constraints are even more restrictive due to the lack of double-buffering for the shift register).

Due to the bit-reversal algorithm being assembler there are some hacks.  CMake support for assembler isn't great.  The XC8 driver / assembler is 'weird'.  The best I could come up with was adding the assembly files into the linker command because I couldn't get the relocatable .objs to generate with the assembler pre-processor enabled.