Removed the UART TX register (CLC3) from the SM16703P modulator - the timing is such that combinational logic alone works.
Re-purposed CLC3 as an on-chip bodge wire due to the IR receiver having UART RX wired to RA7 on the board.  This is now routed via CLC3 to RB0 (previously DEBUG0) where PPS can take over.

Added underscore prefices to the private initialisation functions of the SM16703P modulator, to normalise conventions with the other modules.

Added board-errata.txt to keep track of the hardware issues.